[UNIFICATION, SHELL] Re-arrange shell structure
Refactor the entire shell related stuff to make the bas/zsh experience
more consistent.

There was lot of duplication in the configuration for zsh and bash. So,
this commit will fix it by making both the shells use common *.shrc
files for configuration, residing in shell-common. The shell-specific
files have been moved to shell-* folders.

This commit also breaks the prompt management across multiple files.
These files are modular and can be replaced with some other file to
change the look and behaviour of the prompt.

Renames:
- bash to shell-bash
- zsh to shell-zsh

New folder:
- shell-common