[SHELL] Switch to a different loading mechanism
BREAKING CHANGE: The repo becomes non-functional due to this change
unless things are renamed or such.

Require all the files to end in shrc or the name of the shell if they
are to be loaded on startup. Takes code from @holman's dotfiles, to
determine the loading order. Packages will be ported incrementally.