[MANAGE] Replace setup with check
Checking if the system is set up correctly is more powerful than trying to setup the system automatically. It allows me to selectively apply changes and manually be able to verify the system status as I go about setting up various parts of the system, manually.

Additionally, these checks can also print instructions or link to resources which would be useful whne setting up the system.